+++ package.json
@@ -29,7 +29,7 @@
     "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
     "test": "jest -c jest.config.js --coverage",
     "rollup": "rollup -c rollup.config.js",
-    "rollup-min": "rollup -c rollup.config.js --environment minify"
+    "rollup-min": "rollup -c rollup.min-config.js"
   },
   "dependencies": {
     "@croct-tech/sdk": "0.0.1-beta1"
