+++ rollup.min-config.js
@@ -1,17 +1,21 @@
 import resolve from 'rollup-plugin-node-resolve';
 import typescript from 'rollup-plugin-typescript2';
 import tempDir from 'temp-dir';
-import dts from 'rollup-plugin-dts';
+import {uglify} from 'rollup-plugin-uglify';
 import commonjs from 'rollup-plugin-commonjs';
 
 export default () => {
     return [
         {
-            input: 'src/index.ts',
+            input: 'src/plug.ts',
             output: {
-                file: 'build/index.js',
-                format: 'commonjs',
-                sourcemap: true,
+                file: 'build/plug.min.js',
+                name: 'croct',
+                format: 'iife',
+                sourcemap: false,
+            },
+            treeshake: {
+                propertyReadSideEffects: false,
             },
             plugins: [
                 resolve(),
@@ -20,12 +24,13 @@ export default () => {
                     cacheRoot: `${tempDir}/.rpt2_cache`,
                     useTsconfigDeclarationDir: true,
                 }),
+                uglify({
+                    compress: {
+                        unused: true,
+                        dead_code: true,
+                    },
+                }),
             ],
-        },
-        {
-            input: './build/declarations/index.d.ts',
-            output: [{file: 'build/index.d.ts', format: 'commonjs'}],
-            plugins: [dts({respectExternal: true})],
         },
     ];
 };
